move files, create scaffold for rest of plugin
