--- src/telliot_ampl/__init__.py
@@ -1,1 +1,0 @@
-"""AMPL/USD/VWAP plugin"""
