--- src/telliot_ampl/reporter.py
@@ -1,1 +1,0 @@
-"""AMPL/USD/VWAP Reporter"""
