+++ src/telliot_ampl_feeds/__init__.py
@@ -1,0 +1,1 @@
+"""AMPL/USD/VWAP plugins."""
