+++ src/telliot_ampl_feeds/feeds/__init__.py
@@ -1,0 +1,1 @@
+"""Telliot AMPL feeds"""
