+++ src/telliot_ampl_feeds/reporter.py
@@ -1,0 +1,1 @@
+"""AMPL Reporters"""
