+++ src/telliot_ampl_feeds/scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Scripts for reporting values using AMPL datafeeds."""
